refactor: Session 객체를 상속 받지 않도록 변경
