+++ tomcat/src/main/java/org/apache/coyote/http11/controller/util/BodyExtractor.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package org.apache.coyote.http11.controller.util;
 
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -10,6 +10,9 @@ public class BodyExtractor {
     private static final String INVIDUAL_QUERY_PARAM_DIVIDER = "&";
     private static final String QUERY_PARAM_KEY_VALUE_SPLIT = "=";
     private static final int DONT_HAVE_VALUE = 1;
+
+    private BodyExtractor() {
+    }
 
     public static Map<String, String> convertBody(ResponseBody responseBody) {
         return Stream.of(responseBody.getBody().split(INVIDUAL_QUERY_PARAM_DIVIDER))
